src/backend/app/services/cv_service.py
import cv2
import numpy as np

class CVService:
    def validate_document_image(self, file_path: str):
        """
        Validates if the image has the characteristics of a valid document/passport.
        """
        try:
            # In a real scenario, we would read the image bytes properly
            # For this prototype, we simulate the check or read a dummy file if provided
            
            # Simple check: is it a valid image file?
            # img = cv2.imread(file_path)
            # if img is None:
            #     return {"valid": False, "reason": "Corrupted or invalid image format"}

            # Mock Logic for Prototype success
            return {
                "valid": True,
                "detected_features": ["Face", "MRZ Code", "Passport Layout"],
                "confidence": 0.95
            }
        except Exception as e:
            return {"valid": False, "reason": str(e)}

cv_service = CVService()
