src/backend/pyproject.toml
[project]
name = "youvisa-backend"
version = "0.1.0"
description = "YOUVISA Sprint 2 Backend"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.110.0",
    "uvicorn>=0.29.0",
    "python-multipart>=0.0.9",
    "python-dotenv>=1.0.1",
# AI & NLP
    "openai>=1.14.0",
# Computer Vision
    "opencv-python-headless>=4.9.0.80",
    "numpy>=1.26.4",
# Testing
    "pytest>=8.1.1",
    "httpx>=0.27.0"
]

[tool.setuptools]
packages = ["app"]
